% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_transect.R
\name{crop_transect}
\alias{crop_transect}
\title{Crop transect}
\usage{
crop_transect(transect_obj, map_obj)
}
\arguments{
\item{transect_obj}{Dataframe. Les transects.}

\item{map_obj}{Dataframe. Le tableau avec les données de la carte}
}
\value{
Dataframe. Les transects croppés sur la taille de la grille.
}
\description{
Crop transect
}
\examples{
# TO DO 
}
