% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_obs.R
\docType{data}
\name{dataset_obs}
\alias{dataset_obs}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\author{
Maud Quéroué \email{maud.queroue@gmail.com}
}
\keyword{data}
\keyword{obs}
