% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detection}
\alias{detection}
\title{Detection function}
\usage{
detection(dist_obj, key, esw_km = NA, g_zero = NA, truncation_m)
}
\arguments{
\item{dist_obj}{dataframe. Contains each observation and the distance between the observation and the transects/segments.}

\item{key}{character. Form of the detection function "hn" or "unif".}

\item{esw_km}{numeric. Effective strip width (in km). This argument is needed only for the 'hn' form. By default NA.}

\item{g_zero}{numeric. The detection probability for the 'unif' form. By default NA.}

\item{truncation_m}{numeric. A single numeric value (in m) describing the longest distance at which an object may be observed.}
}
\value{
dataframe. The dist dataframe with new columns : "proba" (numeric values between 0 and 1), the probability of being detection and 'detected' (0 or 1)informing if the individual is detected by the sample design.
}
\description{
This funtion allows simulating the probability that a individual could be observed according to the sample design. The detection function could be an uniform detection function with a probability of detection g_zero on the whole strip band until the maximum distance of observation (in m) truncation_m. The detection function could also be a half normal detection function for which we can choose the effective strip width (in km) esw_km i.e. the distance at which there are as much non detected individuals before this distance than detected individuals after this distance. For the half normal detection function it is also possible to choose the proability of detection at 0 meter g_zero and the maximum distance of observation (in m) truncation_m.
}
\examples{

library(ggplot2)

data(dataset_dist)

detected <- detection(dist_obj = dataset_dist,
                   key = "hn",
                   esw_km = 0.16,
                   g_zero = 1,
                   truncation_m = 400)

ggplot(detected, aes(x=distance_m, y=proba)) +
  geom_point(color = "indianred4") +
  xlim(0,500)


detected <- detection(dist_obj = dataset_dist,
                   key = "unif",
                   g_zero = 0.8,
                   truncation_m = 250) 

ggplot(detected, aes(x=distance_m, y=proba)) +
  geom_point(color = "indianred4") +
  xlim(0,500)

}
