% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esw_hn.R
\name{esw_hn}
\alias{esw_hn}
\title{Calculate the effective strip width of a half normal function}
\usage{
esw_hn(sigma)
}
\arguments{
\item{sigma}{numeric. Sigma of half normal function.}
}
\value{
numeric. esw in km associated to the sigma given.
}
\description{
This function comes from the pelaDSM package. This functions allows to calculate the effective strip width corresponding to the sigma of a half normal detection function.
}
\examples{

esw_hn(sigma = 0.3)

# esw of 0.376 km

}
