% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_map.R
\name{extract_map}
\alias{extract_map}
\title{Extract map with the desired density}
\usage{
extract_map(density_obj, N, crs)
}
\arguments{
\item{density_obj}{Density object from dsims package. The map containing information density ratio on the study area.}

\item{N}{Numeric. The number of individuals desired in the area.}

\item{crs}{Numeric. Projection system.}
}
\value{
sf object. The map with the densities corresponding to the number of individuals desired in the study area.
}
\description{
This function allows, from a density map density_obj (density object of the dsims packages), to provide a map of class sf (data.frame).It allows to recalculate the correct density ratios of the provided map according to the desired number of individuals in the area N.
}
\examples{

data(dataset_density)

map <- extract_map(density_obj = dataset_density,
                   N = 500,
                   crs = 2154)


head(map)

}
