% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monitored_area.R
\name{get_monitored_area}
\alias{get_monitored_area}
\title{Obetnir l'aire de la zone d'étude qui est controlée par les transects}
\usage{
get_monitored_area(transects_data, region_data, truncation_m, crs)
}
\arguments{
\item{transects_data}{dataframe. Les transects}

\item{region_data}{region object. La region d'étude}

\item{truncation_m}{numeric. La distance à laquelle on ne peut plus observer d'animaux. En m.}

\item{crs}{numeric. Le systeme de projection}
}
\value{
numeric. L'aire de la surface d'étude monitorée en m
}
\description{
Obetnir l'aire de la zone d'étude qui est controlée par les transects
}
\examples{
# TO DO 
}
