% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercali-package.R
\docType{package}
\name{intercali-package}
\alias{intercali}
\alias{intercali-package}
\title{intercali: Intercalibration Des Différents Types De Suivis Aeriens}
\description{
Ce package contient de nombreuses fonctions permettant de simuler des données de suivis aérien. A partir de ces simulations de données, il est possible de comparer les estimations d'abondance et de distribution d'espèces.
}
\author{
\strong{Maintainer}: Maud Quéroué \email{maud.queroue@gmail.com}

}
\keyword{internal}
