% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_detect.R
\name{plot_detect}
\alias{plot_detect}
\title{PLot detection}
\usage{
plot_detect(dist_obj, transect_obj, map_obj, title)
}
\arguments{
\item{dist_obj}{sf dataframe. Distances between individuals and associated transect/segment.}

\item{transect_obj}{sf dataframe. Transect/segments data.}

\item{map_obj}{dataframe. La carte de densité.}

\item{title}{character. Le titre souhiaté pour le graphique.}
}
\value{
ggplot object. A map highligthing detected individuals.
}
\description{
PLot detection
}
