% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dsm.R
\name{plot_dsm}
\alias{plot_dsm}
\title{Plot dsm}
\usage{
plot_dsm(
  dsm_pred_data,
  map_data,
  title = "Detection",
  legend = "Marsouins\\n(ind/km2)"
)
}
\arguments{
\item{dsm_pred_data}{dsm object. Le modele dsm créé.}

\item{map_data}{dataframe. La carte avec la densité associée}

\item{title}{character. Le titre souhaité pour le graphique}

\item{legend}{character. La légende souhiatée pour le graphique}
}
\value{
plot. La zone d'étude avec ce que le modèle a predit.
}
\description{
Plot dsm
}
\examples{
# TO DO
}
