% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Plot map}
\usage{
plot_map(map_obj, legend = "Marsouins\\n(ind/km2)", title = "")
}
\arguments{
\item{map_obj}{dataframe. La carte avec la densité associée}

\item{legend}{character. La légende souhiatée pour le graphique}

\item{title}{character. Le titre souhaité pour le graphique}
}
\value{
plot. La zone d'étude avec ce que le modèle a predit.
}
\description{
Plot map
}
\examples{
# TO DO
}
