% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_obs.R
\name{plot_obs}
\alias{plot_obs}
\title{Plot obs}
\usage{
plot_obs(obs_obj, map_obj, legend = "Marsouins\\n(ind/km2)", title = "")
}
\arguments{
\item{obs_obj}{dataframe. Les individus simulés et leur localisation.}

\item{map_obj}{dataframe. La carte avec la densité associée}

\item{legend}{character. La légende souhiatée pour le graphique}

\item{title}{character. Le titre souhaité pour le graphique}
}
\value{
plot. La zone d'étude avec ce que le modèle a predit.
}
\description{
Plot obs
}
\examples{
# TO DO
}
