% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transects.R
\name{plot_transects}
\alias{plot_transects}
\title{Plot transect}
\usage{
plot_transects(transect_obj, map_obj, crs, ifsegs)
}
\arguments{
\item{transect_obj}{dataframe. Le tableau de données avec les transects.}

\item{map_obj}{dataframe. La carte avec les densités}

\item{crs}{numeric. Le systeme de coordonnees utilisé.}

\item{ifsegs}{TRUE/FALSE. Est-ce que les transects sont des segments ?}
}
\value{
plot. La carte de la region avec les transects.
}
\description{
Plot transect
}
\examples{
# TO DO 
}
