% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dsm.R
\name{prepare_dsm}
\alias{prepare_dsm}
\title{Obtenir dans le bon format les fichiers nécessaires aux analyses de distance sampling}
\usage{
prepare_dsm(map_data, dist_data, segs_data)
}
\arguments{
\item{map_data}{dataframe. La carte avec la densité associée}

\item{dist_data}{dataframe. Les distances entre individus et transects}

\item{segs_data}{dataframe. Les différents transects utilisés et leur coordonnées}
}
\value{
List. Les différents objects nécessaire pour faire du distance sampling.
}
\description{
Obtenir dans le bon format les fichiers nécessaires aux analyses de distance sampling
}
\examples{
# TO DO
}
