% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_hn.R
\name{scale_hn}
\alias{scale_hn}
\title{Scale hn}
\usage{
scale_hn(esw, lower = 0.001, upper = 5, w = +Inf)
}
\arguments{
\item{esw}{Numeric. effctive strip width}

\item{lower}{Numeric. lower}

\item{upper}{Numeric. upper}

\item{w}{Numeric. Je ne sais pas}
}
\value{
numeric. Le sigma associé à l'esw dans le cas d'une demi normale
}
\description{
Scale hn
}
