% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentize_transect.R
\name{segmentize_transect}
\alias{segmentize_transect}
\title{Segmentize transects}
\usage{
segmentize_transect(transect_obj, length_m, to = "MULTILINESTRING")
}
\arguments{
\item{transect_obj}{sf dataframe. Transect data.}

\item{length_m}{numeric. Length of the segments desired.}

\item{to}{character. Format desired for the output.}
}
\value{
sf dataframe. Segmentized transect data.
}
\description{
Segmentize transects
}
\examples{

data("dataset_transects")

segs <- segmentize_transect(transect_obj = dataset_transects,
                            length_m = 2000,
                            to = "LINESTRING")

head(segs)

}
