% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_and_calculate.R
\name{sim_and_calculate}
\alias{sim_and_calculate}
\title{Simulate data and calculate dsm}
\usage{
sim_and_calculate(
  density_obj,
  transect_obj,
  N,
  crs,
  key,
  esw_km = NA,
  strip_prob = NA,
  truncation_m
)
}
\arguments{
\item{density_obj}{density object. La carte de densité crée avec le package dsims}

\item{transect_obj}{dataframe. Les transects utilisés.}

\item{N}{numeric. Le nombre d'individus dans la zone d'étude}

\item{crs}{numeric. Le systeme de projection.}

\item{key}{Character. Forme de la fonction de détection "hn" ou "unif"}

\item{esw_km}{numeric. Effective strip width (km). Utile que pour la demi normale, sinon NA. Par défaut NA}

\item{strip_prob}{numeric. Le parametre de la loi uniforme. Par défaut NA.}

\item{truncation_m}{numeric. A partir de quelle distance aucun individu ne peut être détecté.}
}
\value{
list. Une liste avec toutes les informations utiles pour l'intercalibration des données.
}
\description{
Simulate data and calculate dsm
}
