% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ind.R
\name{simulate_ind}
\alias{simulate_ind}
\title{Processus de poisson inhomogène}
\usage{
simulate_ind(map_obj, crs)
}
\arguments{
\item{map_obj}{sf. La carte contenant les information de densité}

\item{crs}{numeric. Le systeme de coordonées utilisé}
}
\value{
dataframe. Les localisations des individus.
}
\description{
Processus de poisson inhomogène
}
\examples{
# TO DO
}
