% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Calculate distance between transects and points}
\usage{
calculate_distance(x, segments, crs)
}
\arguments{
\item{x}{sf object. Les individus simules}

\item{segments}{sf object. Les segments de transects}

\item{crs}{numeric. Le systeme de coordonnes geographiques}
}
\value{
dataframe. Un tableau donnant les distances par rapport aux individus et le segment associé.
}
\description{
Calculate distance between transects and points
}
\examples{
# TO DO
}
