% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_transect.R
\name{create_transect}
\alias{create_transect}
\title{Create transect}
\usage{
create_transect(
  region_obj,
  crs,
  design,
  design.angle,
  line.length,
  truncation,
  ...
)
}
\arguments{
\item{region_obj}{Region object from dssd package. The map containing information about the geometry of the study area.}

\item{crs}{Numeric. Projection system.}

\item{design}{Character. Variable describing the type of design. Either "random", "systematic", "eszigzag" (equal-spaced zigzag), "eszigzagcom" (equal spaced zigzag with complementary lines) or "segmentedgrid". See dssd package for more information.}

\item{design.angle}{Numeric. Value detailing the angle of the design. A value of -1 will cause a random design angle to be generated. See dssd package for more information.}

\item{line.length}{Numeric. The approximative total line length desired.}

\item{truncation}{Numeric. A single numeric value (in m) describing the longest distance at which an object may be observed.}

\item{...}{All other arguments that could be used in the make.design function. See dssd package for more information.}
}
\value{
sf object. The created transects.
}
\description{
Create transect
}
\examples{
library(sf)
library(dssd)
library(dsims)

# Use of the St Andrews bay map from the dssd package
shapefile.name <- system.file("extdata", "StAndrew.shp", package = "dssd")

# Creation of the object with the make.region function of the dsims package
region <- make.region(region.name = "St Andrews bay",
                      shape = shapefile.name,
                      units = "m")

transects <- create_transect(region_obj = region,
                             crs = 2154,
                             design = "eszigzag",
                             line.length = 400000,
                             design.angle = 30,
                             truncation = 400)

head(transects)

transects \%>\%
  st_length() \%>\%
  sum()

}
