% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_transect.R
\name{create_transect}
\alias{create_transect}
\title{Create transect}
\usage{
create_transect(reg, crs, design, spacing, design.angle, truncation)
}
\arguments{
\item{reg}{Formal class region object. La region creer à partir de make.region}

\item{crs}{Numeric.Le systeme de coordonnees geographiques}

\item{design}{a character variable describing the type of design. Either "random", "systematic", "eszigzag" (equal-spaced zigzag), "eszigzagcom" (equal spaced zigzag with complementary lines) or "segmentedgrid". See details for more information. See package dssd pour information}

\item{spacing}{used by systematic designs, numeric value(s) to define spacing between transects. Can be a vector of values with one value per stratum.}

\item{design.angle}{numeric value detailing the angle of the design. Can provide multiple values relating to strata. The use of the angle varies with design, it can be either the angle of the grid of points, the angle of lines or the design axis for the zigzag design. See details. In addition, a value of -1 will cause a random design angle to be generated.}

\item{truncation}{A single numeric value describing the longest distance at which an object may be observed.}
}
\value{
sf object. Les transects crees
}
\description{
Create transect
}
\examples{
# TO DO 
}
