% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_transect.R
\name{crop_transect}
\alias{crop_transect}
\title{Crop transect}
\usage{
crop_transect(transect_obj, map_obj, ifsegs = FALSE)
}
\arguments{
\item{transect_obj}{sf dataframe. The transect data.}

\item{map_obj}{sf dataframe. Map of the study area with the density.}

\item{ifsegs}{Boolean. TRUE to highlight the different segments with colors. By default FALSE.}
}
\value{
sf dataframe. The transect resize to match the density map grid.
}
\description{
Crop transect
}
\examples{

data(dataset_transects)
data(dataset_map)

# The transects do not correspond perfectly to the density map
plot_transects(transect_obj = dataset_transects, 
               map_obj = dataset_map, 
               crs = 2154)

# Crop transects
cropped_transects <- crop_transect(transect_obj = dataset_transects,
                                   map_obj = dataset_map)

# The transects correspond perfectly to the density map
plot_transects(transect_obj = cropped_transects, 
               map_obj = dataset_map, 
               crs = 2154)

}
