% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detection}
\alias{detection}
\title{Fonction de détection}
\usage{
detection(dist_obj, key, esw_km = NA, strip_prob = NA, truncation_m)
}
\arguments{
\item{dist_obj}{Dataframe avec les donnees de distance}

\item{key}{character. Forme de la fonction de détetcion "hn" ou "unif"}

\item{esw_km}{numeric. Effective strip width (km). Utile que pour la demi normale, sinon NA. Par défaut NA.}

\item{strip_prob}{numeric. Le parametre de la loi uniforme. Par défaut NA.}

\item{truncation_m}{numeric. A partir de quelle distance aucun individu ne peut être détecté.}
}
\value{
dataframe. Le tableau des distances avec en plus l'ajout de colonnes de porbabilité de détection et de la détection 1/0 calculée.
}
\description{
Fonction de détection
}
