% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esw_hn.R
\name{esw_hn}
\alias{esw_hn}
\title{esw_hn}
\usage{
esw_hn(sigma)
}
\arguments{
\item{sigma}{numeric. Sigma of half normal function.}
}
\value{
numeric. esw value associated to the sigma given.
}
\description{
esw_hn
}
\examples{

esw_hn(sigma = 0.3)

# esw of 0.376 km

}
