% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esw_hn.R
\name{esw_hn}
\alias{esw_hn}
\title{Esw_hn}
\usage{
esw_hn(sigma, w = +Inf)
}
\arguments{
\item{sigma}{Numeric. Le sigma d'une fonction demi normale}

\item{w}{Numeric. Je ne sais pas}
}
\value{
Numeric. La valeur de esw associée au sigma.
}
\description{
Esw_hn
}
