% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_map.R
\name{extract_map}
\alias{extract_map}
\title{Extract map with the desired density}
\usage{
extract_map(density_obj, N, crs)
}
\arguments{
\item{density_obj}{Density object from dsims package. The map containing information density ratio on the studied region.}

\item{N}{Numeric. The numeber of individuals desired in the area.}

\item{crs}{Numeric. Projection system.}
}
\value{
sf object. The map with the densities corresponding to the number of individuals desired in the studied area.
}
\description{
Extract map with the desired density
}
\examples{

data(dataset_density)

map <- extract_map(density_obj = dataset_density,
                   N = 500,
                   crs = 2154)


head(map)
  
}
