% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_map.R
\name{extract_map}
\alias{extract_map}
\title{Extraire une carte avec les densités souhaitées}
\usage{
extract_map(obj_dens, N, crs)
}
\arguments{
\item{obj_dens}{Density object form dsims package. La carte contenant les densités.}

\item{N}{Numeric. Le nombre d'individus dans la zone d'étude}

\item{crs}{Numeric. Le système de coordonnées utilisé.}
}
\value{
sf object. La carte avec les densité souhaitées.
}
\description{
Extraire une carte avec les densités souhaitées
}
\examples{
# TO DO

}
