% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monitored_area.R
\name{get_monitored_area}
\alias{get_monitored_area}
\title{Obetnir l'aire de la zone d'étude qui est controlée par les transects}
\usage{
get_monitored_area(transect_obj, map_obj, truncation_m)
}
\arguments{
\item{transect_obj}{dataframe. Les transects}

\item{map_obj}{dataframe. La carte de densité}

\item{truncation_m}{numeric. La distance à laquelle on ne peut plus observer d'animaux. En m.}
}
\value{
numeric. L'aire de la surface d'étude monitorée en m
}
\description{
Obetnir l'aire de la zone d'étude qui est controlée par les transects
}
\examples{
# TO DO 
}
