% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monitored_area.R
\name{get_monitored_area}
\alias{get_monitored_area}
\title{Get the area of the monitored region}
\usage{
get_monitored_area(transect_obj, map_obj, truncation_m, crs = NA)
}
\arguments{
\item{transect_obj}{sf dataframe. Transect/segment data.}

\item{map_obj}{sf dataframe or Region object from dssd. Study region.}

\item{truncation_m}{Numeric. A single numeric value (in m) describing the longest distance at which an object may be observed.}

\item{crs}{Numeric. Projection system. Default : NA.}
}
\value{
numeric. The area of the monitored surface in m.
}
\description{
Get the area of the monitored region
}
\examples{
data("dataset_map")
data("dataset_segs")

get_monitored_area(transect_obj = dataset_segs,
                   map_obj = dataset_map,
                   truncation_m = 400)
}
