% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_dsm.R
\name{get_var_dsm}
\alias{get_var_dsm}
\title{Recuperer des estimations de variance}
\usage{
get_var_dsm(grid_obj, dsm_obj)
}
\arguments{
\item{grid_obj}{dataframe. La grille de prédiction du modele}

\item{dsm_obj}{dsm objet. Le modele dsm créé}
}
\value{
List. Une liste avec le cv, le se et l'intervalle de confiance à 95\%
}
\description{
Recuperer des estimations de variance
}
\examples{
# TO DO
}
