% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_detect.R
\name{plot_detect}
\alias{plot_detect}
\title{PLot detection}
\usage{
plot_detect(obs_data, dist_data, transect_data, region_data, crs, title)
}
\arguments{
\item{obs_data}{dataframe. Le tableau des données simulées}

\item{dist_data}{dataframe. Le tableau contenant les distance de chaque individu simulé au transects.}

\item{transect_data}{dataframe. Le tableau renseignant sur les transects.}

\item{region_data}{region object. l'objet région crée avec le package dsims.}

\item{crs}{numeric. Le systeme de projection.}

\item{title}{character. Le titre souhiaté pour le graphique.}
}
\value{
plot. Une carte mettant en évidence les individus détectés.
}
\description{
PLot detection
}
