% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_detect.R
\name{plot_detect}
\alias{plot_detect}
\title{PLot detection}
\usage{
plot_detect(dist_obj, transect_obj, map_obj, title)
}
\arguments{
\item{dist_obj}{dataframe. Le tableau contenant les distance de chaque individu simulé au transects.}

\item{transect_obj}{dataframe. Le tableau renseignant sur les transects.}

\item{map_obj}{dataframe. La carte de densité.}

\item{title}{character. Le titre souhiaté pour le graphique.}
}
\value{
plot. Une carte mettant en évidence les individus détectés.
}
\description{
PLot detection
}
