% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_detect.R
\name{plot_detect}
\alias{plot_detect}
\title{Plot detection}
\usage{
plot_detect(dist_obj, transect_obj, map_obj, title)
}
\arguments{
\item{dist_obj}{sf dataframe. Distances between individuals and associated transect/segment and probability that the individual is detected or not.}

\item{transect_obj}{sf dataframe. Transect/segments data.}

\item{map_obj}{sf dataframe. Map of the study area with the density.}

\item{title}{character. Title for the graph.}
}
\value{
ggplot object. A map highligthing detected individuals.
}
\description{
Plot detection
}
\examples{
data("dataset_detected")
data("dataset_segs")

plot_detect(dist_obj = dataset_detected, 
            transect_obj = dataset_segs, 
            map_obj = dataset_map, 
            title = "Detected individuals")
}
