% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Plot map}
\usage{
plot_map(map_obj, legend = "Density (ind/km2)", title = "")
}
\arguments{
\item{map_obj}{Dataframe. Sf map to plot.}

\item{legend}{Character. Legend for the color gradient.}

\item{title}{Character. Title.}
}
\value{
ggplot object. The study area with the gradient of density (ind/km2).
}
\description{
This fonction allows to plot the map created with the extract_map function. It is nessary to use the map_obj, a sf dataframe, contaning at least, a column density_km. The title and the legend can be personnalized.
}
\examples{

data(dataset_density)

map <- extract_map(density_obj = dataset_density,
                   N = 500,
                   crs = 2154)

plot_map(map_obj = map)

}
