% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_obs.R
\name{plot_obs}
\alias{plot_obs}
\title{Plot simulated individuals}
\usage{
plot_obs(obs_obj, map_obj, legend = "Density (ind/km2)", title = "")
}
\arguments{
\item{obs_obj}{sf dataframe. Individuals simulated with their coordinates.}

\item{map_obj}{sf dataframe. Map of the study area with a density column density_km.}

\item{legend}{Character. Legend.}

\item{title}{Character. Title.}
}
\value{
ggplot object. Study area with the simulates individuals and the density color gradient.
}
\description{
Plot simulated individuals
}
\examples{

data(dataset_map)

ind <- simulate_ind(map_obj = dataset_map,
                    crs = 2154)

plot_obs(obs_obj = ind,
         map_obj = dataset_map)

}
