% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transects.R
\name{plot_transects}
\alias{plot_transects}
\title{Plot transect}
\usage{
plot_transects(transect_obj, map_obj, crs, ifsegs = FALSE)
}
\arguments{
\item{transect_obj}{sf dataframe. Transect data.}

\item{map_obj}{sf dataframe or Region object from dssd. Study region.}

\item{crs}{Numeric. Projection system}

\item{ifsegs}{Boolean. TRUE to highlight the different segments with colors. By default FALSE.}
}
\value{
ggplot object. The transects represented on the study area.
}
\description{
Plot transect
}
\examples{

library(dssd)
library(dsims)
data(dataset_transects)
data(dataset_segs)
data(dataset_map)


# Use of the St Andrews bay map from the dssd package
shapefile.name <- system.file("extdata", "StAndrew.shp", package = "dssd")

# Creation of the object with the make.region function of the dsims package
region <- make.region(region.name = "St Andrews bay",
                      shape = shapefile.name,
                      units = "m")

# Plot transects
plot_transects(transect_obj = dataset_transects, 
               map_obj = region, 
               crs = 2154)

# Vizualize segment
plot_transects(transect_obj = dataset_segs, 
               map_obj = dataset_map, 
               crs = 2154,
               ifsegs = TRUE)

}
