% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_density.R
\name{random_density}
\alias{random_density}
\title{Create a random density}
\usage{
random_density(
  region_obj,
  grid_m,
  density_base,
  crs,
  amplitude,
  sigma,
  nb_simu
)
}
\arguments{
\item{region_obj}{region object. The region create with dsims}

\item{grid_m}{numeric. La taille des grilles qui sont des carré (coté du carré en m)}

\item{density_base}{numeric. La densité de base pour la carte}

\item{crs}{numeric. Le systeme de coordonnées géographique}

\item{amplitude}{numeric. Le min et le max de l'amplitude des hotspots ajoutés}

\item{sigma}{numeric. Le min et le max de l'étendue des hotspots ajoutés}

\item{nb_simu}{numeric. Nombre de hotspot a ajouter}
}
\value{
density object. La carte avec les densités crées
}
\description{
Create a random density
}
