% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentize_transect.R
\name{segmentize_transect}
\alias{segmentize_transect}
\title{Title}
\usage{
segmentize_transect(x, to = "MULTILINESTRING")
}
\arguments{
\item{x}{Fichier sf Polygon ou Linestring. Les transects qu'ils faut découper en segments}

\item{to}{Type de fichier desire en sortie}
}
\value{
Le fichier sf avec les segments
}
\description{
Title
}
\examples{
# TO DO 
}
