% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ind.R
\name{simulate_ind}
\alias{simulate_ind}
\title{Simulate individuals with a inhomogenous Poisson point process}
\usage{
simulate_ind(map_obj, crs)
}
\arguments{
\item{map_obj}{Dataframe. Sf map with a colum containg density informations density_m}

\item{crs}{Numeric. Projection system.}
}
\value{
Dataframe. Indivduals with their coordinates associated.
}
\description{
From an sf class density map \code{map_obj} (data.frame), an inohomogene Poisson point process is used to simulate the presence of individuals in the study area. The probability of presence of an individual is dependent on the density given by the map.
}
\examples{

data(dataset_map)

ind <- simulate_ind(map_obj = dataset_map,
                    crs = 2154)

head(ind)

}
